# codr-api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Users demo
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /api
...
schemes:
  - http
  - https
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
paths:
...
  /users:
    x-swagger-router-controller: users
    get:
      description: Returns a list of all user accounts.
      operationId: getUsers
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserList"
        "401":
          description: Not allowed
        "500":
          description: Internal server error
  /users/current:
    x-swagger-router-controller: users
    get:
      description: Returns information for the currently logged in user.
      operationId: getCurrentUser
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CurrentUser"
        "401":
          description: Not allowed
...
  /users/{user_id}:
    x-swagger-router-controller: users
    get:
      description: Returns the details of the specified user account.
      operationId: getUserById
      parameters:
        - name: user_id
          in: path
          description: ID of customer to fetch account for
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/CurrentUser"
        "401":
          description: Not allowed
        "404":
          description: User_Id not found
        "405":
          description: Incorrect input
        "500":
          description: Internal server error
    delete:
      description: Deletes the specified user account.
      operationId: deleteUserById
      parameters:
        - name: user_id
          in: path
          description: ID of customer to fetch account for
          required: true
          type: integer
          format: int64
      responses:
        "204":
          description: Item deleted
        "401":
          description: Not allowed
        "404":
          description: Id not found
        "405":
          description: Incorrect input
        "500":
          description: Internal server error
    put:
      description: Updates the specified user account.
      operationId: putUserById
      parameters:
        - name: user_id
          in: path
          description: ID of customer to fetch account for
          required: true
          type: integer
          format: int64
        - name: body
          in: body
          description: body of the request
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/User"
        "401":
          description: Not allowed
        "404":
          description: Id not found
        "405":
          description: Incorrect input
        "500":
          description: Internal server error
  /users/{user_id}/comments:
    x-swagger-router-controller: users
    get:
      description: Returns all comments for the specified user.
      operationId: getUserCommentsById
      parameters:
        - name: user_id
          in: path
          description: ID of customer to fetch account for
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserCommentResponseList"
        "405":
          description: Incorrect input
        "500":
          description: Internal server error
...
definitions:
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      display_name:
        type: string
  CurrentUser:
    type: object
    properties:
      id:
        type: integer
        format: int64
      user_level:
        type: integer
        format: int64
      display_name:
        type: string
  UserList:
      type: array
      items:
        $ref: '#/definitions/User'
...
  UserCommentResponse:
    type: object
    properties:
      id:
        type: integer
        format: int64
      parent_id:
        type: integer
        format: int64
      file_id:
        type: integer
        format: int64
      project_id:
        type: integer
        format: int64
      exercise_id:
        type: integer
        format: int64
      course_id:
        type: integer
        format: int64
      module_id:
        type: integer
        format: int64
      line_range:
        type: string
      content:
        type: string
      author_id:
        type: integer
        format: int64
      visible:
        type: boolean
      timestamp:
        type: integer
        format: int64
      deleted:
        type: boolean
      can_edit:
        type: boolean
      can_delete:
        type: boolean
      can_toggle_visibility:
        type: boolean
  UserCommentResponseList:
      type: array
      items:
        $ref: '#/definitions/UserCommentResponse'
...